requirements.txt
discord.py~=2.4.0
openai~=1.51.2
python-dotenv~=1.0.1
requests~=2.32.3
